\name{Shinypkg-package}
\alias{Shinypkg-package}
\alias{Shinypkg}
\docType{package}
\title{
\packageTitle{Shinypkg}
}
\description{
\packageDescription{Shinypkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Shinypkg}
\packageIndices{Shinypkg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Shinypkg}

Maintainer: \packageMaintainer{Shinypkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
