% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{Shinypkg vote information}
\alias{Shinypkg vote information}
\title{vote information of parliament members}
\usage{
shinyApp(ui = ui, server = server)
}
\value{
ui. Sidebar with a slider input for approve votes
}
\description{
take the data from APILab and APIpkg and ploting vote frequency of the number of approve for each parties member.
}
